+++ app/models/eve/system.rb
@@ -4,7 +4,7 @@ module Eve
   class System < ApplicationRecord
     # TODO: belongs_to :constellation_id
 
-    has_many :stargates, primary_key: :system_id
+    has_many :stargates, primary_key: :system_id # rubocop:disable Rails/HasManyOrHasOneDependent
 
     # TODO: belongs_to :star_id
   end